CIP-2067 - undoing change

+++ src/ska_mid_cbf_mcs/subarray/subarray_component_manager.py
@@ -660,7 +660,7 @@ class CbfSubarrayComponentManager(
                     t = Thread(
                         target=self._update_timing_beam_weights,
                         args=(
-                            float(timing_beam_weights["epoch"]),
+                            int(timing_beam_weights["epoch"]),
                             json.dumps(
                                 timing_beam_weights[
                                     "timing_beam_weights_details"
